+++ backend/Repositories/IGraphRepository.cs
@@ -8,7 +8,7 @@ public interface IGraphRepository
     Task<Graph?> GetBySlugAsync(string slug, CancellationToken cancellationToken = default);
     Task<bool> DeleteNodeAsync(string slug, string nodeId, CancellationToken cancellationToken = default);
     Task<bool> AddNodeAsync(string slug, GraphNodeDto node, string? parentID = null,
-        string edgeKind = "support", int importanceToParent = 1, CancellationToken cancellationToken = default);
+        string edgeKind = "support", decimal importanceToParent = 1m, CancellationToken cancellationToken = default);
     Task<bool> UpdateNodeAsync(string slug, string nodeId, GraphNodeUpdateDto node,
         CancellationToken cancellationToken = default);
     Task<bool> AddEdgeAsync(string slug, GraphEdgeDto edge, CancellationToken cancellationToken = default);
